order command implementation

+++ ITService.Domain/Command/Employee/AddEmployeeCommand.cs
@@ -7,7 +7,7 @@ namespace ITService.Domain.Command.Employee
         public string Login { get; set; }
         public string Password { get; set; }
         public string Email { get; set; }
-        public int? Salary { get; set; }
+        public decimal? Salary { get; set; }
         public Guid? RoleId { get; set; }
     }
 }
