+++ ITService.Domain/Command/Order/AddOrderCommand.cs
@@ -6,7 +6,7 @@ namespace ITService.Domain.Command.Order
     {
         public Guid? UserId { get; set; }
         public Guid? EmployeeId { get; set; }
-        public int? Amount { get; set; }
+        public decimal? Amount { get; set; }
         public string OrderStatus { get; set; }
         public DateTime? OrderDate { get; set; }
     }
